Separate GameProgress and GameProgressModal markup
